requirements.txt
google-cloud-language
bokeh
pandas
tqdm